Fixes to make batch.sh run correctly
The last sed command in this script was creating a malformed json. It
was stripping the trailing single-quote of the json string but not the
leading one.

Changed the sed expression to remove single quotes from the json string.
Also moved adding the "--config" option to the place where parallel is
invoking valhalla_run_route. Since the arg to this option is fixed for
all runs of parallel, I find it easier to read if placed in the same
place where the executable is being called. This also limits the sed
expression to whats needed.

Tested by running batch.sh as follows:
$ ./batch.sh ../test_requests/demo_routes.txt ../build/valhalla.json